requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
langchain==0.2.6
langchain-groq==0.1.6
langchain-community==0.2.6
langchain-experimental==0.0.62
langsmith==0.1.83
sqlalchemy==2.0.31
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
python-dotenv==1.0.1
pydantic==2.7.4
requests==2.32.3
tabulate==0.9.0